+++ core/data/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/data/artemis_context/ArtemisContextProviderImpl.kt
@@ -1,8 +1,6 @@
-package de.tum.informatics.www1.artemis.native_app.core.data.artemis_context.impl
+package de.tum.informatics.www1.artemis.native_app.core.data.artemis_context
 
 import android.util.Log
-import de.tum.informatics.www1.artemis.native_app.core.data.artemis_context.ArtemisContext
-import de.tum.informatics.www1.artemis.native_app.core.data.artemis_context.ArtemisContextProvider
 import de.tum.informatics.www1.artemis.native_app.core.data.service.network.AccountDataService
 import de.tum.informatics.www1.artemis.native_app.core.datastore.AccountService
 import de.tum.informatics.www1.artemis.native_app.core.datastore.ServerConfigurationService
